refactor: split auto-loop.ts monolith into auto/ directory modules
auto-loop.ts was 1892 lines — a single file with natural section boundaries
marked by ─── delimiters. Split into 7 focused modules under auto/:

- auto/types.ts       — constants (MAX_LOOP_ITERATIONS, BUDGET_THRESHOLDS)
                        and all shared types (AgentEndEvent, UnitResult,
                        PhaseResult, IterationContext, LoopState, etc.)
- auto/resolve.ts     — mutable resolve state + cross-module setters
                        (_setCurrentResolve, _setSessionSwitchInFlight,
                        _clearCurrentResolve) + public API
- auto/detect-stuck.ts — detectStuck() (leaf)
- auto/loop-deps.ts   — LoopDeps interface (leaf, type-only)
- auto/run-unit.ts    — runUnit()
- auto/phases.ts      — all 5 pipeline phases (runPreDispatch, runDispatch,
                        runGuards, runUnitPhase, runFinalize) + helpers
- auto/loop.ts        — autoLoop()

auto-loop.ts becomes a ~16-line barrel re-exporting the public API.
Zero behavior changes. All external consumers unchanged.

Also update 5 structural test files that were readFileSync-ing auto-loop.ts
to now point at the correct submodule where each pattern lives.

1826/1826 tests pass.
