+++ src/resources/extensions/gsd/auto/types.ts
@@ -1,0 +1,99 @@
+/**
+ * auto/types.ts — Constants and types shared across auto-loop modules.
+ *
+ * Leaf node in the import DAG — no imports from auto/.
+ */
+
+import type { ExtensionAPI, ExtensionContext } from "@gsd/pi-coding-agent";
+
+import type { AutoSession } from "./session.js";
+import type { GSDPreferences } from "../preferences.js";
+import type { GSDState } from "../types.js";
+import type { CmuxLogLevel } from "../../cmux/index.js";
+import type { LoopDeps } from "./loop-deps.js";
+
+/**
+ * Maximum total loop iterations before forced stop. Prevents runaway loops
+ * when units alternate IDs (bypassing the same-unit stuck detector).
+ * A milestone with 20 slices × 5 tasks × 3 phases ≈ 300 units. 500 gives
+ * generous headroom including retries and sidecar work.
+ */
+export const MAX_LOOP_ITERATIONS = 500;
+/** Maximum characters of failure/crash context included in recovery prompts. */
+export const MAX_RECOVERY_CHARS = 50_000;
+
+/** Data-driven budget threshold notifications (descending). The 100% entry
+ *  triggers special enforcement logic (halt/pause/warn); sub-100 entries fire
+ *  a simple notification. */
+export const BUDGET_THRESHOLDS: Array<{
+  pct: number;
+  label: string;
+  notifyLevel: "info" | "warning" | "error";
+  cmuxLevel: "progress" | "warning" | "error";
+}> = [
+  { pct: 100, label: "Budget ceiling reached", notifyLevel: "error", cmuxLevel: "error" },
+  { pct: 90, label: "Budget 90%", notifyLevel: "warning", cmuxLevel: "warning" },
+  { pct: 80, label: "Approaching budget ceiling — 80%", notifyLevel: "warning", cmuxLevel: "warning" },
+  { pct: 75, label: "Budget 75%", notifyLevel: "info", cmuxLevel: "progress" },
+];
+
+// ─── Types ───────────────────────────────────────────────────────────────────
+
+/**
+ * Minimal shape of the event parameter from pi.on("agent_end", ...).
+ * The full event has more fields, but the loop only needs messages.
+ */
+export interface AgentEndEvent {
+  messages: unknown[];
+}
+
+/**
+ * Result of a single unit execution (one iteration of the loop).
+ */
+export interface UnitResult {
+  status: "completed" | "cancelled" | "error";
+  event?: AgentEndEvent;
+}
+
+// ─── Phase pipeline types ────────────────────────────────────────────────────
+
+export type PhaseResult<T = void> =
+  | { action: "continue" }
+  | { action: "break"; reason: string }
+  | { action: "next"; data: T }
+
+export interface IterationContext {
+  ctx: ExtensionContext;
+  pi: ExtensionAPI;
+  s: AutoSession;
+  deps: LoopDeps;
+  prefs: GSDPreferences | undefined;
+  iteration: number;
+}
+
+export interface LoopState {
+  recentUnits: Array<{ key: string; error?: string }>;
+  stuckRecoveryAttempts: number;
+}
+
+export interface PreDispatchData {
+  state: GSDState;
+  mid: string;
+  midTitle: string;
+}
+
+export interface IterationData {
+  unitType: string;
+  unitId: string;
+  prompt: string;
+  finalPrompt: string;
+  pauseAfterUatDispatch: boolean;
+  observabilityIssues: unknown[];
+  state: GSDState;
+  mid: string | undefined;
+  midTitle: string | undefined;
+  isRetry: boolean;
+  previousTier: string | undefined;
+}
+
+export type WindowEntry = { key: string; error?: string };
